--- grazer/src/main/java/edu/asu/diging/grazer/core/document/impl/IDocument.java
@@ -1,5 +1,0 @@
-package edu.asu.diging.grazer.core.document.impl;
-
-public interface IDocument {
-
-}
